requirements.txt
accelerate==1.0.1
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
dnspython==2.6.1
faiss-cpu==1.8.0.post1
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
fsspec==2025.3.0
ftfy==6.2.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mongomock==4.3.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime==1.19.2
open_clip_torch==2.32.0
packaging==24.2
pillow==10.4.0
protobuf==5.29.4
psutil==7.0.0
pymongo==4.10.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentinels==1.0.0
sympy==1.13.3
timm==1.0.15
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2

